add separate entity for local Repo model, add DBFlow, add integrations

+++ app/src/main/java/es/guillermoorellana/travisforandroid/api/entity/ApiRepo.java
@@ -29,7 +29,7 @@ public class ApiRepo implements Serializable {
     private String slug;
     private boolean active;
     private String description;
-    private int lastBuildId;
+    private long lastBuildId;
     private String lastBuildNumber;
     private String lastBuildState;
     private long lastBuildDuration;
